passed the previously aadded failing test using the fileApi which is use to read the file and clientapi which is use to call the rest api to get the single bitcoin vaalue

+++ src/test/java/portfolio/BitcoinPortfolioTest.java
@@ -17,7 +17,7 @@ public class BitcoinPortfolioTest {
         final var totalValueOfEachBitcoin = bitcoinPortfolio.calculateBitcoinTotalValue(bitcoinPortfolio.getFileApi().getLines());
 
         assertEquals(1, totalValueOfEachBitcoin.size());
-        assertTrue(totalValueOfEachBitcoin.get("BTC") != 0);
+        assertTrue(totalValueOfEachBitcoin.get("BTC").doubleValue()!=0);
     }
 
 }
